pyproject.toml
[project]
name = "fast-api-acumen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "azure-identity>=1.23.0",
    "azure-keyvault-secrets>=4.9.0",
    "azure-monitor-opentelemetry>=1.6.9",
    "azure-storage-blob>=12.25.1",
    "coverage>=7.8.0",
    "fastapi[standard]>=0.115.12",
    "httpx[http2]>=0.28.1",
    "loguru>=0.7.3",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.9.1",
    "pydantic[email]>=2.11.4",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-cov>=6.1.1",
    "python-jose[cryptography]>=3.4.0",
    "sqlalchemy>=1.4",
]
